test: add coverage for travis component

+++ _Utilities/county-specific-test-case-builder.ps1
@@ -1,6 +1,36 @@
 ﻿<#
     county specific test case builder
 #>
+
+function getBlankTestContent($csname){
+    $arr = @(
+    "using LegalLead.PublicData.Search.Classes;",
+    "using LegalLead.PublicData.Search.Util;",
+    "using Moq;",
+    "using OpenQA.Selenium;",
+    "using System;",
+    "",
+    "namespace legallead.search.tests.util",
+    "{",
+    "    public class ~0",
+    "    {",
+    "         [Fact]",
+    "         public void ServiceHasTypeDefined()",
+    "         {",
+    "             var error = Record.Exception(() => _ = typeof(~1));",
+    "             Assert.Null(error);",
+    "         }",
+    "    }",
+    "}"
+    );
+    [string]$cmmd = [string]::Join( [System.Environment]::NewLine, $arr )
+    $tpname = $csname.Replace( "Tests", "" );
+    $cmmd = $cmmd.Replace( "~0", $csname );
+    $cmmd = $cmmd.Replace( "~1", $tpname );
+    return $cmmd;
+}
+
+
 $dest = "Travis";
 $src = "Dallas";
 $pattern = "*$dest*.cs";
@@ -14,6 +44,7 @@ $found | ForEach-Object {
     $templateName = [string]::Concat( [System.IO.Path]::GetFileNameWithoutExtension( $info.Name.Replace($dest, $src) ), "Tests.cs");
     $testFile = [System.IO.Path]::Combine( $targetdir, $name );
     $templateFile = [System.IO.Path]::Combine( $targetdir, $templateName );
+    
     if ([System.IO.File]::Exists( $templateFile ) -eq $true -and [System.IO.File]::Exists( $testFile ) -eq $false ) {
         $content = [System.IO.File]::ReadAllText( $templateFile );
         $incidentId = $content.IndexOf( $src );
@@ -21,6 +52,11 @@ $found | ForEach-Object {
             $content = $content.Replace( $src, $dest );
             $incidentId = $content.IndexOf( $src );
         }
+        [System.IO.File]::WriteAllText( $testFile, $content )
+    }
+    if ([System.IO.File]::Exists( $testFile ) -eq $false ) {
+        $cname = [System.IO.Path]::GetFileNameWithoutExtension( $testFile )
+        $content = (getBlankTestContent -csname $cname)
         [System.IO.File]::WriteAllText( $testFile, $content )
     }
 }
